DONTMERGE: (blind trial for travis build)
The update to INSTALL shown here is bogus, but was made only to show
where necessary information to make use of the new feature needs to
be added.

Specifically, the patches by Jean-Noel Avila requires po4a even for
those who do not want localized documentation, and what is described
in INSTALL pretends that this has already been fixed.

Signed-off-by: Junio C Hamano <gitster@pobox.com>